fix: bug on sentiment reading frontmatter
