+++ src/statefield-plugin.ts
@@ -8,6 +8,7 @@ import { Summary, updateSummary } from "./bridge";
 import { ObsidianReadabilitySettings } from "./settings";
 import { PLUGINS } from "./retext-plugins";
 import matter from "gray-matter";
+import { getFrontmatterRange } from "./utils";
 
 type Classes = Record<
   string,
@@ -58,21 +59,6 @@ const initializeProcessor = (settings: ObsidianReadabilitySettings) => {
   return processor.use(stringify);
 };
 
-const getFrontmatterRange = (text: string) => {
-  text = text.trim();
-  const begin = text.indexOf("---");
-  let end = text.indexOf("---", begin + 3);
-
-  while (end !== -1 && text[end - 1] !== "\n" && text[end + 4] !== "\n") {
-    end = text.indexOf("---", end + 3);
-  }
-
-  if (end === -1 && begin === 0) return [begin, text.length];
-  else if (begin !== 0 || end === -1) return [0, 0];
-
-  return [begin, end];
-};
-
 export const errorHighlightPlugin = (settings: ObsidianReadabilitySettings) =>
   StateField.define<DecorationSet>({
     create: (_) => Decoration.none,
@@ -93,8 +79,6 @@ export const errorHighlightPlugin = (settings: ObsidianReadabilitySettings) =>
         const begin = message.position?.start.offset || 0;
         const end = message.position?.end.offset || 0;
         const isInsideFrontMatter = begin >= beginFm && end <= endFm;
-
-        console.log(beginFm, endFm, begin, end, isInsideFrontMatter);
 
         if (!message.source || isInsideFrontMatter) continue;
         const source = message.source;
